backend/requirements.txt

# backend\requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
Flask-Mail==0.9.1
werkzeug==3.0.1
